chore(server): rename field metadata service v2
